requirements.txt
openai-whisper
pyannote.audio
librosa
tqdm
python-dotenv
termcolor
pydub
SpeechRecognition